+++ app/main/views/sign_in.py
@@ -1,10 +1,12 @@
 from flask import render_template, redirect, jsonify
-from flask_login import login_user
+from flask import session
 
 from app.main import main
 from app.main.dao import users_dao
 from app.main.encryption import checkpw
+from app.main.encryption import hashpw
 from app.main.forms import LoginForm
+from app.main.views import send_sms_code
 
 
 @main.route("/sign-in", methods=(['GET']))
@@ -24,7 +26,9 @@ def process_sign_in():
         if not user.is_active():
             return jsonify(active_user=False), 401
         if checkpw(form.password.data, user.password):
-            login_user(user)
+            sms_code = send_sms_code(user.mobile_number)
+            session['user_id'] = user.id
+            session['sms_code'] = hashpw(sms_code)
         else:
             users_dao.increment_failed_login_count(user.id)
             return jsonify(authorization=False), 401
